BT: uint16_t numHandles

+++ cpp_utils/BLEService.h
@@ -27,7 +27,7 @@ public:
 	void setByUUID(BLECharacteristic* pCharacteristic, const char* uuid);
 	void setByUUID(BLECharacteristic* pCharacteristic, BLEUUID uuid);
 	void setByHandle(uint16_t handle, BLECharacteristic* pCharacteristic);
-	BLECharacteristic* getByUUID(const char* uuid);	
+	BLECharacteristic* getByUUID(const char* uuid);
 	BLECharacteristic* getByUUID(BLEUUID uuid);
 	BLECharacteristic* getByHandle(uint16_t handle);
 	BLECharacteristic* getFirst();
@@ -69,8 +69,8 @@ public:
 	uint8_t			   m_id = 0;
 
 private:
-	BLEService(const char* uuid, uint32_t numHandles);
-	BLEService(BLEUUID uuid, uint32_t numHandles);
+	BLEService(const char* uuid, uint16_t numHandles);
+	BLEService(BLEUUID uuid, uint16_t numHandles);
 	friend class BLEServer;
 	friend class BLEServiceMap;
 	friend class BLEDescriptor;
@@ -88,7 +88,7 @@ private:
 	FreeRTOS::Semaphore  m_semaphoreStartEvt  = FreeRTOS::Semaphore("StartEvt");
 	FreeRTOS::Semaphore  m_semaphoreStopEvt   = FreeRTOS::Semaphore("StopEvt");
 
-	uint32_t             m_numHandles;
+	uint16_t             m_numHandles;
 
 	BLECharacteristic* getLastCreatedCharacteristic();
 	void               handleGATTServerEvent(
